refactor all python code WIP
