+++ autonomy/nodes/arc_turn_manual_node.py
@@ -1,0 +1,5 @@
+#!/usr/bin/python3
+from autonomy.automodule.arc_turn_manual import main
+
+if __name__ == "__main__":
+    main()
