+++ autonomy/nodes/autonomy_node.py
@@ -1,0 +1,5 @@
+#!/usr/bin/python3
+from autonomy.automodule.autonomy import main
+
+if __name__ == "__main__":
+    main()
