+++ autonomy/nodes/robot_state_node.py
@@ -1,0 +1,5 @@
+#!/usr/bin/python3
+from autonomy.automodule.robot_state import main
+
+if __name__ == "__main__":
+    main()
