+++ autonomy/setup.py
@@ -7,7 +7,10 @@ from catkin_pkg.python_setup import generate_distutils_setup
 
 # fetch values from package.xml
 setup_args = generate_distutils_setup(
-    packages=['uwb_localization'],
+    packages=['autonomy',
+              'autonomy.automodule',
+              'autonomy.path_following',
+              'autonomy.path_planning'],
     package_dir={'': 'src'},
 )
