+++ autonomy/src/autonomy/path_following/skid_steer_simulator.py
