+++ glenn_launcher/launch/autonomy/autonomy.launch
@@ -16,6 +16,6 @@
         <node machine="robot" name="robot_state_node" pkg="autonomy" type="robot_state.py" output="screen"/>
 
         <node name="autonomy_node" pkg="autonomy" type="autonomy_node.py" launch-prefix="xterm -e"/>
-	    <node name="visualizer_node" pkg="autonomy" type="visualizer.py"/>
+	    <node name="visualizer_node" pkg="autonomy" type="visualizer_node.py"/>
     </group>
 </launch>
