+++ glenn_launcher/launch/autonomy/autonomy_interface.launch
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
 <launch>
 	<node name="autonomy_node" pkg="autonomy" type="autonomy_node.py" launch-prefix="xterm -e"/>
-	<node name="visualizer_node" pkg="autonomy" type="visualizer.py"/>
+	<node name="visualizer_node" pkg="autonomy" type="visualizer_node.py"/>
 </launch>
