+++ glenn_launcher/launch/autonomy/simulator.launch
@@ -5,6 +5,6 @@
     <node name="transit_node" pkg="autonomy" type="transit_node.py" launch-prefix="xterm -e"/>
     <node name="autonomy_node" pkg="autonomy" type="autonomy_node.py" launch-prefix="xterm -e"/>
     <node name="robot_state_node" pkg="autonomy" type="robot_state.py" output="screen"/>
-    <node name="simulator_node" pkg="autonomy" type="simulator.py"/>
-    <node name="visualizer_node" pkg="autonomy" type="visualizer.py"/>
+    <node name="simulator_node" pkg="autonomy" type="simulator_node.py"/>
+    <node name="visualizer_node" pkg="autonomy" type="visualizer_node.py"/>
 </launch>
