+++ obstacle_detection/setup.py
@@ -1,10 +1,14 @@
-## ! DO NOT MANUALLY INVOKE THIS setup.py, USE CATKIN INSTEAD 
-from distutils.core import setup 
-from catkin_pkg.python_setup import generate_distutils_setup 
- 
-# fetch values from package.xml 
-setup_args = generate_distutils_setup( 
-packages=['obstacle_detection'], 
-package_dir={'': 'include'}, 
-) 
+#!/usr/bin/env python
+
+# ! DO NOT MANUALLY INVOKE THIS setup.py, USE CATKIN INSTEAD
+
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+# fetch values from package.xml
+setup_args = generate_distutils_setup(
+    packages=['obstacle_detection'],
+    package_dir={'': 'src'},
+)
+
 setup(**setup_args)
